[rtems] Add an init task and configs for RTEMS
The RTEMS init task is recognised by the RTEMS framework
to be the entry point for RTEMS applications. This init
task here is simple and just calls coremark's main.

RTEMS Config macros are provided to setup an environment
that coremark needs such as a timer, console driver, (optional) threads, etc