posix: Fix SEPARATE_COMPILE for derived ports
Trying to build with PORT_DIR=freebsd SEPARATE_COMPILE=1 fails as it
tries to build freebsd/core_portme.o from freebsd/core_portme.c.
Arguably we should be building posix/core_portme.o, but that is awkward
to implement due to part of the logic existing in the root Makefile as
we'd need to make it know to mkdir posix/ instead of freebsd/ (for the
OPATH != ./ case). Instead we can just add another rule in the posix
core_portme.mak that maps PORT_DIR to posix.